S0C-3A-1A: Add shadow_verify_dual_write_sampling scenario and update registry and documentation
